FIX #3040 Slow test
